higher level rendering, especially lines
wrote funcs to plot 3 points and lines onto the film.
refactored main render loop accordingly.
removed the "scale" param in favor of moving the camera.
added parts to get and draw the "cube" of the simulation bounds.

apparently I had misunderstood what mgl was doing in some of the
transform coord functions, especially in regards to things being
behind the camera. this lead to the weird "wrap" visual bug I had
seen in the point clouds earlier. Clipping points was easy.
I had to write bits to perform correction and clipping for lines
that extended behind the camera, which was harder and required a
lot of research and experimentation.